requirements.txt
terminal-shop==1.6.0
python-dotenv==1.1.0
pynput==1.8.1